draw 9 boxes and give them the right material values.

+++ 05_GLBasicLight/box_shader.frag
@@ -9,7 +9,7 @@ struct Material {
 	vec3 ambient;
 	vec3 diffuse;
 	vec3 specular;
-	int shininess;
+	float shininess;
 };
 
 struct Light{
